+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Updated declaration file extension so it will not be included in final build
 - Added multiple date format support for `expiry` field
+- Added `dayjs` package
 
 ## 3.0.1 (July 11, 2021)